Source code of the LOCFIT software system
LOCFIT is a software system for fitting curves and surfaces to data,
using the local regression and likelihood methods. The code is mostly
written in C; an interface is provided that enables LOCFIT to be used as
an S-Plus or R library.

This is the version dated August 1, 2001 and available for download on
the following page:

    http://cm.bell-labs.com/cm/ms/departments/sia/project/locfit/

All credit is due to the authors, who have kindly released this code
for free use and distribution (subject to conditions in the project
README).

- Catherine Loader
  catherine@research.bell-labs.com or cathl@optonline.net
  http://cm.bell-labs.com/stat/project/locfit/
  Copyright (c) 1996-2001 Lucent Technologies, Bell Laboratories.

- Also incorporating SCB code, joint with Jiayang Sun
  jiayang@sun.stat.cwru.edu
  http://sun.cwru.edu/~jiayang/sci3.html
  SCB code is Copyright (c) 1996-2001 Jiayang Sun.